move TestAssessmentHTMLView into its own file
